+++ main/Support_Functions.cpp
@@ -1,0 +1,38 @@
+#include"Category.h"
+
+//--------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------
+
+bool Check_Category(const char* Path)
+{
+    struct stat info;
+
+    if (stat(Path, &info) != 0)
+        return false;
+    else if (info.st_mode & S_IFDIR)
+        return true;
+    else
+        return false;
+}
+
+
+
+//--------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------
+
+int confirm(char Input)                              // SUPPORT FUNCTION FOR ACTION CONFIRMATIONS
+{
+    if (&Input == "Y")
+    {
+        return 1;
+    }
+    else if (&Input == "N")
+    {
+        return 2;
+    }
+    else
+    {
+        return 3;
+    }
+}
+
+
+//--------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------
